Split agent tool answer formatters
