+++ Tests/QuillCodeParityTests/ParityAgentGateTests.swift
@@ -4,13 +4,16 @@ final class ParityAgentGateTests: QuillCodeParityTestCase {
     func testAgentRunnerDelegatesFinalAnswerFormatting() throws {
         let agentText = try Self.agentSourceText(named: "Agent.swift")
         let builderText = try Self.agentSourceText(named: "AgentFinalAnswerBuilder.swift")
+        let formatterText = try Self.agentSourceText(named: "AgentToolAnswerFormatters.swift")
 
         XCTAssertTrue(builderText.contains("enum AgentFinalAnswerBuilder"), "Tool-result final answer copy should live in a focused builder.")
         XCTAssertTrue(builderText.contains("static func finalAnswer"), "Final answer formatting should be directly testable.")
-        XCTAssertTrue(builderText.contains("private static var toolAnswerFormatters"), "Tool-specific final answers should be registered instead of growing the top-level finalAnswer method.")
-        XCTAssertTrue(builderText.contains("for formatter in toolAnswerFormatters"), "The top-level final-answer method should route through the formatter registry.")
-        XCTAssertTrue(builderText.contains("ToolDefinition.shellRun.name"), "Shell final-answer special cases should live in the builder.")
-        XCTAssertTrue(builderText.contains("ToolDefinition.browserInspect.name"), "Browser final-answer special cases should live in the builder.")
+        XCTAssertTrue(builderText.contains("AgentToolAnswerFormatters.all"), "The top-level final-answer method should route through the formatter registry.")
+        XCTAssertFalse(builderText.contains("import QuillCodeTools"), "The builder should not import tool modules once formatting has a separate owner.")
+        XCTAssertTrue(formatterText.contains("enum AgentToolAnswerFormatters"), "Tool-specific final-answer summaries should live in a dedicated formatter owner.")
+        XCTAssertTrue(formatterText.contains("static var all: [Formatter]"), "Tool-specific final answers should be registered instead of growing the top-level finalAnswer method.")
+        XCTAssertTrue(formatterText.contains("ToolDefinition.shellRun.name"), "Shell final-answer special cases should live with tool formatters.")
+        XCTAssertTrue(formatterText.contains("ToolDefinition.browserInspect.name"), "Browser final-answer special cases should live with tool formatters.")
         XCTAssertTrue(agentText.contains("AgentFinalAnswerBuilder.finalAnswer"), "AgentRunner should delegate final-answer formatting.")
         XCTAssertFalse(agentText.contains("private static func shellAnswer"), "AgentRunner should not own shell final-answer formatting.")
         XCTAssertFalse(agentText.contains("private static func browserInspectionAnswer"), "AgentRunner should not own browser final-answer formatting.")
